Move validate_plugin() and validate_plugin_ids() to plugins.py
