+++ .github/scripts/plugins.py
@@ -162,6 +162,22 @@ def get_core_plugins():
         md_file.write(new_contents)
 
 
+def validate_plugin(plugin, manifest, repo, file_groups):
+    return validate_plugin_ids(plugin, manifest, repo, file_groups)
+
+
+def validate_plugin_ids(plugin, manifest, repo, file_groups):
+    ids_match = True
+    releases_id = plugin.get('id')
+    manifest_id = manifest.get('id')
+    if releases_id != manifest_id:
+        print(
+            f"ERROR repo:{repo} ID {releases_id} does not match ID in manifest: {manifest_id}")
+        file_groups.setdefault("error", list()).append(f"{releases_id}/{manifest_id}")
+        ids_match = False
+    return ids_match
+
+
 def main(argv=sys.argv[1:]):
     get_core_plugins()
